src/requirements.txt
pillow
numpy
gdk
matplotlib
opencv-contrib-python
PyGObject
pynput
pytesseract
scikit-image
scipy
tesseract